third_party/blink/web_tests/wpt_internal/html/canvas/drawElementImage/privacy/svg-mask-image-ignored.https.sub.html

<!DOCTYPE HTML>
<html>
<head>
  <title>drawElementImage does not use cross-origin images in SVG masks</title>
  <link rel="help" href="https://github.com/WICG/html-in-canvas">
  <script src='/resources/testharness.js'></script>
  <script src='/resources/testharnessreport.js'></script>
  <script src='../support/getPixel.js'></script>
  <script src="/wpt_internal/resources/canvas-draw-element/waitForCanvasPaint.js"></script>
</head>
<body>
  <canvas id="canvas" width="100" height="200" layoutsubtree>
    <svg width="100" height="200">
      <defs>
        <mask id="mask-same" maskContentUnits="objectBoundingBox">
          <image href="https://{{location[host]}}/wpt_internal/html/canvas/drawElementImage/resources/blue-100x50-transparent-100x50.png" x="0" y="0" width="1" height="1" preserveAspectRatio="none" />
        </mask>
        <mask id="mask-cross" maskContentUnits="objectBoundingBox">
          <image href="https://{{hosts[alt][www]}}:{{ports[https][0]}}/wpt_internal/html/canvas/drawElementImage/resources/blue-100x50-transparent-100x50.png" x="0" y="0" width="1" height="1" preserveAspectRatio="none" />
        </mask>
      </defs>
      <rect x="0" y="0" width="100" height="100" fill="blue" mask="url(#mask-same)"/>
      <rect x="0" y="100" width="100" height="100" fill="blue" mask="url(#mask-cross)"/>
    </svg>
  </canvas>

  <script>
    window.onload = () => {
      promise_test(async function(t) {
        await waitForCanvasPaint(canvas);
        var ctx = canvas.getContext('2d');
        ctx.drawElementImage(document.querySelector('canvas > svg'), 0, 0);

        // Fetch all pixel data once to avoid multiple slow readbacks.
        const imgData = ctx.getImageData(0, 0, canvas.width, canvas.height);

        let pixel = getPixel(imgData, 50, 25);
        assert_array_equals(pixel, [0, 0, 0, 0], "Same origin mask should fully mask top-half of rect");

        pixel = getPixel(imgData, 50, 75);
        assert_array_equals(pixel, [0, 0, 255, 18], "Same origin mask should not mask bottom-half of rect");

        pixel = getPixel(imgData, 50, 125);
        assert_array_equals(pixel, [0, 0, 0, 0], "Cross origin mask data should not affect top-half of rect");

        pixel = getPixel(imgData, 50, 175);
        assert_array_equals(pixel, [0, 0, 0, 0], "Cross origin mask data should not affect bottom-half of rect");
      });
    }
  </script>
</body>
</html>
